app/app/pyproject.toml
[tool.poetry]
name = "Mushi-identifier-app"
version = "0.1.0"
description = ""
authors = ["jpusa <49599941+jpusa@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
tensorflow = "^2.6.0"
numpy = "^1.19.2"
uvicorn = "^0.15.0"
gunicorn = "^20.1.0"
fastapi = "^0.68.1"
Pillow = "^8.3.2"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
